refactor(compare): extract crosswalk to crosswalk.py (Phase A 4/N)
Move ~1970 lines (load_phv_*, type-dispatch helpers, YAML parsing,
expected-summary building, _aggregate_source_summaries,
build_variable_crosswalk, and ~30 supporting private helpers) into
hv_dataqc/compare/crosswalk.py. This is the largest single extraction
in Phase A.

The new file imports CrosswalkBuildError from _common and uses
canonical_phv_id from hv_dataqc_common (aliased as _canonical_phv_id
inside the module to preserve call sites verbatim).

Check functions still need a few crosswalk-private helpers
(_distribution_count_map, _codes_are_numeric_or_sentinel, _normalize_code,
_is_null_sentinel_code). These are imported alongside the public API.

Tests import many crosswalk-private symbols by name from
compare_source_harmonized. To keep test imports working without a
broader test rewrite, those symbols are re-exported from
compare_source_harmonized via a `from hv_dataqc.compare.crosswalk import ...`
block — flagged with `noqa: F401` because the linter can't tell that
re-exports are intentional. The test imports will move directly to
hv_dataqc.compare.crosswalk in a later cleanup.

The crosswalk module is 2022 lines — over the 500-line per-file target
in the plan, but the recon flagged that crosswalk is a cohesive unit
better kept whole than split. Splitting it (e.g., crosswalk_cache.py
for cache loaders) is a possible follow-up.

Regression-check: COPDGene, ARIC, HCHS — MD + JSON all match. Tests:
79/79 pass. --help: works.

Main file is now 1733 lines, down from ~4090 at the start of Phase A.
